+++ src/python/lib/workflowUtil.py
@@ -371,12 +371,14 @@ def cleanPyEnv() :
     clear out some potentially destabilizing env variables:
     """
 
-    # Stopping default clearing of python env variables (MANTA-1317)
+    # Stopping default clearing of python env variables (MANTA-1316)
     # There are cases where module'd-in pythons will not operate
     # after this change. The motivation for clearing were cases where
     # a user PYTHONPATH library interferes with a workflow function, but
     # if these are encountered in the future, we should have a more
     # specfic diagnostic/solution to the problem
+    #
+    # Discussion in the context of manta here: https://github.com/Illumina/manta/issues/116
     #
     clearList = [] # [ "PYTHONPATH", "PYTHONHOME"]
     for key in clearList :